Add daily goals.
Add an optional `daily_goals` block for the `report` configuration of an activity type.

Currently this block has only one supported attribute: `distance_km`, whose value is a float.

When posting a daily report to slack, indicate if the day's distance total has exceeded the daily `distance_km` goal for that activity type.