Implement streak reporting.
For this iteration, streaks are only reported for distance.

If a goal is set for distance, the streak is the last X days, including today, where the distance goal was met.

If no goal was set for distance, the streak is the last X days, including today, where there was activity.

The streak information is included as part of the "distance" item in the Slack message.