Adjust test to reproduce the but with an incorrect streak calculation.
The bug happens if we did at least once, in the past, reach our goal, but not today. The bug: we took the last day we met our goal, and counted the number of days since then, as the streak.

This wasn't caught by tests before, because the test scenario didn't have a case where a previous activity met the goal.

This commit makes a test fail.

The following commit fixes it.