Avoid duplicate Slack messages.
Sometimes, fitbit or withings notify us multiple times, for the same event, within the same second or so.

We already have some logic inside the notification webhook routes to avoid processing duplicate events:
1 Check if the event was marked as processed.
2 If marked as processed, exit.
3 Process the event.
4 Mark the event as processed.

The problem is: During "process the event", we get another call on our webhook. We haven't processed it, so we haven't yet marked it as processed. So, it gets processed twice, and the Slack message gets sent twice.

This commit only addresses the slack message duplication.

We need to think more about avoiding processing the duplicated events in the first place.
* If we mark the event as processed before actually starting to process it, if an exception occurs during the processing, the next duplicate event becomes a missed opportunity: we could have been able to try to process it again, but we won't.
* We could add a mutex (asyncio Lock) around the object that holds the info about the event being processed or not, but this might make our route slower to respond.

We'll (maybe) think about handling this more cleanly at the source (in the webhook notification), but for now, this commit just prevents duplicate slack messages being posted.