Refactor activity streak calculation, keeping current streak calculation logic.
Make the repository return the number of days in the streak, instead of the earliest date in the streak.

The future streak calculation logic will evolve: it will ignore days with no activity at all. So returning an "earliest date in the streak" won't help to calculate the number days in the streak.